per-fishery results table, manually ported
